backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0
djangorestframework-simplejwt>=5.0,<6.0
dj-database-url>=3.0,<4.0
gunicorn>=25.0,<26.0
whitenoise>=6.0,<7.0
psycopg2-binary>=2.9,<3.0
